ai-api/requirements.txt
llama-cpp-python
flask
fask_cors
